função lista
