+++ src/services/firebaseConfig.js
@@ -6,5 +6,5 @@ const firebaseConfig = {
   messagingSenderId: "418935778179",
   appId: "1:418935778179:web:7486d8e6441f290e17e4b6"
 };
-
 firebase.initializeApp(firebaseConfig);
+
