+++ src/services/index.js
@@ -24,3 +24,4 @@ export const signInGoogle = () => {
 export const signOut = () => firebase.auth().signOut();
 
 export const checkLogin = () => firebase.auth().onAuthStateChanged((user) => user);
+
